+++ infra/grafana/folders.tf
@@ -1,0 +1,6 @@
+resource "grafana_folder" "tcgscan" {
+  provider = grafana.cloud
+
+  title = "TCG Scan"
+  uid   = "tcgscan"
+}